sample metrics code

+++ metrics/fid_score.py
@@ -247,7 +247,7 @@ def calculate_fid_given_paths(paths, batch_size, cuda, dims, bootstrap=True, n_b
         if os.path.isdir(p):
             pths.append(p)
         elif p.endswith('.npy'):
-            np_imgs = np.load(p)
+            np_imgs = np.load(p).astype(np.float64)
             if np_imgs.shape[0] > 25000:
                 np_imgs = np_imgs[:50000]
             pths.append(np_imgs)
